Remove NetBIOSName support from machine_name
Due to issues with setting this value (and presumably because we are not
dealing with domain joining at this point), support for NetBIOSName and
the plist modification is being removed.

Not only was this value causing idempotence and convergence issues, but
the changes were not reflected in the GUI and sometimes even the plist
itself after it was modified and had been through testing.

Signed-off-by: Eric Hanko <v-erhank@microsoft.com>