Minor modifications to keep awake
The only significant change is to make the test suite match
the default attribute of not disabling disk sleep. I also removed some
tests that are not relevant to any modifications we're making to the
power settings.

It should also be noted that the plist we're checking for powernap is
different for El Capitan, so I went ahead and removed that test for now
(it was skipped previously)

Signed-off-by: Eric Hanko <v-erhank@microsoft.com>