Add an attribute to manage VNC screen lock setting
When using Apple Remote Desktop to control computers, disconnecting from
a machine will automatically bring it into a 'sleep' state - this causes
the screen to lock automatically given the screen lock setting in System
Preferences -> Security & Privacy, which is sometimes a hindrance when
needing to observe the state of the machine regularly.
This attribute allows consumers to manage that setting.